Fix tests for wicket 10
