+++ instrumentation/wicket-8.0/common-testing/src/main/java/hello/ExceptionPage.java
@@ -8,6 +8,8 @@ package hello;
 import org.apache.wicket.markup.html.WebPage;
 
 public class ExceptionPage extends WebPage {
+  private static final long serialVersionUID = 1L;
+
   public ExceptionPage() throws Exception {
     throw new Exception("test exception");
   }
