+++ instrumentation/wicket-8.0/common-testing/src/main/java/hello/HelloPage.java
@@ -9,6 +9,8 @@ import org.apache.wicket.markup.html.WebPage;
 import org.apache.wicket.markup.html.basic.Label;
 
 public class HelloPage extends WebPage {
+  private static final long serialVersionUID = 1L;
+
   public HelloPage() {
     add(new Label("message", "Hello World!"));
   }
