+++ instrumentation/wicket-8.0/javaagent/build.gradle.kts
@@ -15,21 +15,4 @@ dependencies {
   bootstrap(project(":instrumentation:servlet:servlet-common:bootstrap"))
 
   library("org.apache.wicket:wicket:8.0.0")
-
-  testImplementation(project(":testing-common"))
-  testImplementation("org.jsoup:jsoup:1.13.1")
-  testImplementation("org.eclipse.jetty:jetty-server:8.0.0.v20110901")
-  testImplementation("org.eclipse.jetty:jetty-servlet:8.0.0.v20110901")
-
-  testInstrumentation(project(":instrumentation:servlet:servlet-3.0:javaagent"))
-  testInstrumentation(project(":instrumentation:servlet:servlet-javax-common:javaagent"))
-}
-
-val latestDepTest = findProperty("testLatestDeps") as Boolean
-
-// Wicket 9 requires Java 11
-if (latestDepTest) {
-  otelJava {
-    minJavaVersionSupported.set(JavaVersion.VERSION_11)
-  }
 }
